requirements.txt
streamlit
boto3
pypdf
faiss-cpu
openai
langchain
langchain-community
python-dotenv
altair